Move E2E infrastructure out of conftest.py into E2E test file
Per kunal-vaishnavi feedback: conftest.py should not contain E2E-specific
helpers. Moved _get_e2e_test_pkgs_path, _preload_and_init_e2e, and
e2e_manager fixture into test_live_audio_transcription_e2e.py.
